Implemented multi-tenancy singe domain

+++ app/Http/Controllers/UserController.php
@@ -2,71 +2,73 @@
 
 namespace App\Http\Controllers;
 
+use App\Constants\ApplicationPermissionConstants;
+use App\Constants\ApplicationRoleConstants;
 use App\Http\Requests\UserRequest;
+use App\Models\Role;
 use App\Models\User;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
-    /**
-     * Display a listing of the users
-     *
-     * @param  \App\User  $model
-     * @return \Illuminate\View\View
-     */
-    public function index(User $model)
+    public function index()
     {
-        $users = User::paginate(25);
-
+        if (!Gate::allows(ApplicationPermissionConstants::viewUsers)) {
+            abort(401);
+        }
+        $users = Auth::user()->company->users;
         return view('users.index', compact('users'));
     }
 
-    /**
-     * Show the form for creating a new user
-     *
-     * @return \Illuminate\View\View
-     */
     public function create()
     {
-        return view('users.create');
+        if (!Gate::allows(ApplicationPermissionConstants::createUser)) {
+            abort(401);
+        }
+        $roles = Role::whereNot('name', ApplicationRoleConstants::SuperAdmin)->get();
+        return view('users.create', compact('roles'));
     }
 
-    /**
-     * Store a newly created user in storage
-     *
-     * @param  \App\Http\Requests\UserRequest  $request
-     * @param  \App\User  $model
-     * @return \Illuminate\Http\RedirectResponse
-     */
     public function store(UserRequest $request)
     {
+        if (!Gate::allows(ApplicationPermissionConstants::createUser)) {
+            abort(401);
+        }
         $request->merge(['password' => Hash::make($request->get('password'))]);
+        $request->merge(['company_id' => Auth::user()->company->id]);
 
-        User::create($request->all());
+        $id = DB::connection('application')->table('users')->insertGetId([
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => $request->password,
+            'company_id' => $request->company_id,
+            "created_at" => Carbon::now()
+        ]);
+
+        $user = User::find($id);
+
+        $user->assignRole(Role::findById($request->get('user_role_id')));
 
         return redirect()->route('users.index')->withStatus('User successfully created.');
     }
 
-    /**
-     * Show the form for editing the specified user
-     *
-     * @param  \App\User  $user
-     * @return \Illuminate\View\View
-     */
     public function edit(User $user)
     {
+        if (!Gate::allows(ApplicationPermissionConstants::updateUser)) {
+            abort(401);
+        }
         return view('users.edit', compact('user'));
     }
 
-    /**
-     * Update the specified user in storage
-     *
-     * @param  \App\Http\Requests\UserRequest  $request
-     * @param  \App\User  $user
-     * @return \Illuminate\Http\RedirectResponse
-     */
     public function update(UserRequest $request, User $user)
     {
+        if (!Gate::allows(ApplicationPermissionConstants::updateUser)) {
+            abort(401);
+        }
         $hasPassword = $request->get('password');
 
         $request->merge(['password' => Hash::make($request->get('password'))]);
@@ -78,14 +80,11 @@ class UserController extends Controller
         return redirect()->route('users.index')->withStatus('User successfully updated.');
     }
 
-    /**
-     * Remove the specified user from storage
-     *
-     * @param  \App\User  $user
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function destroy(User  $user)
+    public function destroy(User $user)
     {
+        if (!Gate::allows(ApplicationPermissionConstants::deleteUser)) {
+            abort(401);
+        }
         $user->delete();
 
         return redirect()->route('users.index')->withStatus('User successfully deleted.');
